centrifugo/requirements.txt
Django==3.2.12
PyJWT==2.4.0
django-cors-headers==3.13.0
requests==2.28.1